// src/interface/weeklyreport/type.ts
export interface IGetAttendanceParams {
  params: IGetAttendanceContentParams;
}

export interface IGetAttendanceContentParams {
  args: string[];
}
// ...
export interface WeeklyReportState {
  id?: number;
  date?: string;
  employee_code?: string;
  total_work_time?: number;
  shift_start?: number;
  shift_end?: number;
  total_shift_work_time?: number;
  holiday_start?: number;
  holiday_end?: number;
  rest_start?: number;
  rest_end?: number;
  employee_name?: string;
  shift_name?: string;
  explanation?: boolean;
  is_explanation?: boolean;
  type_shifts?: number;
  on_time?: boolean;
  wrong_time?: boolean;
  rest_shifts?: boolean;
  empty_shifts?: boolean;
  extra_hour?: number;
  approval_hour?: number;
  company?: string;
  department?: string;
  job_title?: string;
  attendance_attempt_1?: string;
  attendance_attempt_2?: string;
  attendance_attempt_3?: string;
  attendance_attempt_4?: string;
  attendance_attempt_5?: string;
  attendance_attempt_6?: string;
  attendance_attempt_7?: string;
  attendance_attempt_8?: string;
  attendance_attempt_9?: string;
  attendance_attempt_10?: string;
  attendance_attempt_11?: string;
  attendance_attempt_12?: string;
  attendance_attempt_13?: string;
  attendance_attempt_14?: string;
  attendance_attempt_15?: string;
  last_attendance_attempt?: string;
  time_keeping_code?: string;
  resource_calendar: number;
  missing_checkin_break: boolean;
  split_shift: boolean;
}

export interface DataWeeklyType {
  key: React.ReactNode;
  company?: string;
  department?: string;
  employee_name?: string;
  code?: string;
  children?: DataWeeklyType[];
  rest_shifts?: boolean;
  day_0: any;
  day_1: any;
  day_2: any;
  day_3: any;
  day_4: any;
  day_5: any;
  day_6: any;
}